chore: rename Post folder to Publication (#221)
* chore: rename folder

* chore: rename ViewPost to ViewPublication
