+++ src/components/Publication/Actions/Collect/index.tsx
