+++ src/components/Publication/Actions/Menu/Delete.tsx
