+++ src/components/Publication/FullPost.tsx
