+++ src/components/Publication/HiddenPost.tsx
