+++ src/components/Publication/NewPost/index.tsx
