+++ src/components/Publication/RelevantPeople.tsx
