+++ src/components/Publication/Type/Collected.tsx
