+++ src/components/Publication/Type/CommunityPost.tsx
