+++ src/components/Publication/Type/Mirrored.tsx
