+++ src/components/Publication/Type/index.tsx
