+++ src/components/Report/index.tsx
@@ -1,7 +1,7 @@
 import { gql, useMutation, useQuery } from '@apollo/client'
 import { GridItemEight, GridItemFour, GridLayout } from '@components/GridLayout'
-import { POST_QUERY } from '@components/Post'
-import SinglePublication from '@components/Post/SinglePublication'
+import { POST_QUERY } from '@components/Publication'
+import SinglePublication from '@components/Publication/SinglePublication'
 import SettingsHelper from '@components/Shared/SettingsHelper'
 import PostShimmer from '@components/Shared/Shimmer/PostShimmer'
 import { Button } from '@components/UI/Button'
