+++ src/components/Shared/Navbar/index.tsx
@@ -17,7 +17,7 @@ import MoreNavItems from './MoreNavItems'
 import Search from './Search'
 
 const StaffBar = dynamic(() => import('./StaffBar'))
-const NewPostModal = dynamic(() => import('../../Post/NewPost/Modal'))
+const NewPostModal = dynamic(() => import('../../Publication/NewPost/Modal'))
 
 const PING_QUERY = gql`
   query Ping {
